dependency calculation issues
