--- apps/bar/config/config.exs
@@ -1,5 +1,0 @@
-IO.puts "\n+++++++++++++++++++++++++++++"
-IO.puts "++ bar"
-IO.puts "++"
-
-use Mix.Config
