--- apps/foo/.gitignore
@@ -1,5 +1,0 @@
-/_build
-/cover
-/deps
-erl_crash.dump
-*.ez
