--- apps/foo/config/config.exs
@@ -1,5 +1,0 @@
-IO.puts "\n+++++++++++++++++++++++++++++"
-IO.puts "++ foo"
-IO.puts "++"
-
-use Mix.Config
