+++ apps/helper/README.md
@@ -1,4 +1,4 @@
-# Bar
+# Dependency.Helper
 
 **TODO: Add description**
 
@@ -6,15 +6,15 @@
 
 If [available in Hex](https://hex.pm/docs/publish), the package can be installed as:
 
-  1. Add bar to your list of dependencies in `mix.exs`:
+  1. Add helper to your list of dependencies in `mix.exs`:
 
         def deps do
-          [{:bar, "~> 0.0.1"}]
+          [{:helper, "~> 0.0.1"}]
         end
 
-  2. Ensure bar is started before your application:
+  2. Ensure helper is started before your application:
 
         def application do
-          [applications: [:bar]]
+          [applications: [:helper]]
         end
