+++ apps/helper/lib/helper.ex
@@ -1,0 +1,2 @@
+defmodule Dependency.Helper do
+end
