+++ apps/master/lib/master.ex
@@ -1,0 +1,2 @@
+defmodule Dependency.Master do
+end
