+++ apps/test_only/README.md
@@ -1,0 +1,20 @@
+# Dependency.TestOnly
+
+**TODO: Add description**
+
+## Installation
+
+If [available in Hex](https://hex.pm/docs/publish), the package can be installed as:
+
+  1. Add test_only to your list of dependencies in `mix.exs`:
+
+        def deps do
+          [{:test_only, "~> 0.0.1"}]
+        end
+
+  2. Ensure test_only is started before your application:
+
+        def application do
+          [applications: [:test_only]]
+        end
+
