+++ apps/test_only/lib/test_only.ex
@@ -1,0 +1,2 @@
+defmodule Dependency.TestOnly do
+end
