+++ apps/test_only/mix.exs
@@ -1,14 +1,10 @@
-defmodule Bar.Mixfile do
+defmodule Dependency.TestOnly.Mixfile do
   use Mix.Project
 
   def project do
-    [app: :bar,
+    [app: :test_only,
      version: "0.0.1",
-     build_path: "../../_build",
-     config_path: "../../config/config.exs",
-     deps_path: "../../deps",
-     lockfile: "../../mix.lock",
-     elixir: "~> 1.2-dev",
+     elixir: "~> 1.2",
      build_embedded: Mix.env == :prod,
      start_permanent: Mix.env == :prod,
      deps: deps]
@@ -18,8 +14,7 @@ defmodule Bar.Mixfile do
   #
   # Type "mix help compile.app" for more information
   def application do
-    [applications: [:logger],
-     mod: {Bar, []}]
+    [applications: [:logger]]
   end
 
   # Dependencies can be Hex packages:
@@ -29,10 +24,6 @@ defmodule Bar.Mixfile do
   # Or git/path repositories:
   #
   #   {:mydep, git: "https://github.com/elixir-lang/mydep.git", tag: "0.1.0"}
-  #
-  # To depend on another app inside the umbrella:
-  #
-  #   {:myapp, in_umbrella: true}
   #
   # Type "mix help deps" for more examples and options
   defp deps do
