+++ apps/test_only/test/test_helper.exs
@@ -1,0 +1,1 @@
+ExUnit.start()
