+++ apps/test_only/test/test_only_test.exs
@@ -1,0 +1,8 @@
+defmodule Dependency.TestOnlyTest do
+  use ExUnit.Case
+  doctest Dependency.TestOnly
+
+  test "the truth" do
+    assert 1 + 1 == 2
+  end
+end
