+++ config/config.exs
@@ -1,7 +1,3 @@
-IO.puts "\n+++++++++++++++++++++++++++++"
-IO.puts "++ config: project"
-IO.puts "++"
-
 use Mix.Config
 
 import_config "../apps/*/config/config.exs"
